Massive refactoring to integrate ffmpeg.
ffmpeg has a lot of advantages for us.  Namely, it gives us a lot of
additional flexibility over the encoding, significantly reducing the
headache involved in configuring these parameters.  It also offers us an
RTP/RTCP library implemented directly in libavformat, eliminating the
need to use jlibrtp (which I've hated from the start).

Today, the ffmpeg integration is only working as a proof-of-concept.
There seems to be some issue with frame delay (mplayer is claiming it
has to guess the framerate at 10fps, instead of the configured 24fps).
But the remaining issues should be much easier to resolve now that we
have a well-used library in place here.