Added user alert / subject warning system.
This system is designed as part of an optional policy where the
camcorder node will try to protect the privacy of potential subjects in
the camera field of view.  This is accomplished by playing a noise when
the camcorder is about to active and flashing the LED twice.  Meanwhile,
3 seconds of "please wait" frame will be sent over RTP until the real
camera stream begins sending.

Today this is not configurable but it was coded in such a way that would
make this easy.  Ideally we could introduce the notion of a camcorder or
network "policy" that defines it's operating mode.